[refactor] Refactor interview start logic
Refactor foldering of interview start logic

Related to: #5
