+++ feature/interview/src/main/java/com/tdd/interview/start/InterviewScreen.kt
@@ -1,4 +1,4 @@
-package com.tdd.interview
+package com.tdd.interview.start
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
