--- feature/interview/src/main/java/com/tdd/interview/start/InterviewStartViewModel.kt
@@ -1,4 +1,0 @@
-package com.tdd.interview.start
-
-class InterviewStartViewModel {
-}
