+++ feature/interview/src/main/java/com/tdd/interview/start/ready/InterviewStartViewModel.kt
@@ -1,0 +1,4 @@
+package com.tdd.interview.start.ready
+
+class InterviewStartViewModel {
+}
